test: 修正长整型应用ID测试

+++ weixin-java-cp/src/test/java/me/chanjar/weixin/cp/api/impl/WxCpAgentServiceImplTest.java
@@ -65,7 +65,7 @@ public class WxCpAgentServiceImplTest {
     this.wxCpService.getAgentService().set(WxCpAgent.builder()
       .description("abcddd")
       .logoMediaId("aaaaaaaaaaaaaa")
-      .agentId(agentId.intValue())
+      .agentId(agentId)
       .build());
   }
