+++ weixin-java-cp/src/test/java/me/chanjar/weixin/cp/corpgroup/service/impl/WxCpLinkedCorpServiceImplTest.java
@@ -56,7 +56,7 @@ public class WxCpLinkedCorpServiceImplTest {
   public WxCpCorpGroupConfigStorage wxCpCorpGroupConfigStorage() {
     wxCpCorpGroupConfigStorage = new WxCpCorpGroupDefaultConfigImpl();
     wxCpCorpGroupConfigStorage.setCorpId(wxCpService.getWxCpConfigStorage().getCorpId());
-    wxCpCorpGroupConfigStorage.setAgentId(wxCpService.getWxCpConfigStorage().getAgentId().intValue());
+    wxCpCorpGroupConfigStorage.setAgentId(wxCpService.getWxCpConfigStorage().getAgentId());
     return wxCpCorpGroupConfigStorage;
   }
